Milestone 8: RBAC with role-based guards and resource ownership enforcement

+++ src/laptops/laptops.service.ts
@@ -1,4 +1,4 @@
-import { Injectable,NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable,NotFoundException } from '@nestjs/common';
 import { LaptopInterface } from './interfaces/Laptops.interface';
 import { CreateLaptopDto } from './dto/create-laptop.dto';
 import { updateLaptopDto } from './dto/update-laptop.dto';
@@ -34,8 +34,11 @@ export class LaptopsService {
 
 
   // Adds the new Laptop into the Laptops data
-  async createLaptop(newLaptop: CreateLaptopDto): Promise<Laptop> {
-    const newlapt = await this.laptopRepository.create(newLaptop);
+  async createLaptop(newLaptop: CreateLaptopDto, user:any): Promise<Laptop> {
+    const newlapt = await this.laptopRepository.create({
+      ...newLaptop,
+      userId: user.userId,
+    });
 
     return await this.laptopRepository.save(newlapt);
   }
@@ -44,6 +47,7 @@ export class LaptopsService {
   async updateLaptopById(
     id: number,
     updatedLaptop: updateLaptopDto,
+    user: any,
   ): Promise<Laptop> {
 
     const laptop = await this.laptopRepository.findOne({
@@ -54,21 +58,27 @@ export class LaptopsService {
       throw new NotFoundException(`Laptop with id ${id} not found`);
     }
 
-    // Merge the new incoming data onto the existing database object
+    if (laptop.userId !== user.userId){
+      throw new ForbiddenException('You can only edit your own laptops')
+    }
+    // To merge the incoming updated data with the existing laptop data
     Object.assign(laptop, updatedLaptop);
 
-    // TypeORM sees the ID already exists, so it runs an SQL UPDATE
     return await this.laptopRepository.save(laptop);
   }
 
 
-  async deleteLaptopById(id: number): Promise<Laptop> {
+  async deleteLaptopById(id: number, user: any): Promise<Laptop> {
     const laptop = await this.laptopRepository.findOne({
       where: {id}
     })
 
     if (!laptop){
       throw new NotFoundException(`Laptop with ${id} not found`);
+    }
+
+    if (laptop.userId != user.userId){
+      throw new ForbiddenException('You can only delete you own Laptops');
     }
     return await this.laptopRepository.remove(laptop);
